+++ README.md
@@ -7,12 +7,12 @@
 <!-- ![PyPI - Python Version][pyversionbadge] -->
 [![GitHub][licensebadge]][licenseurl]
 
-<!-- [versionbadge]: https://img.shields.io/pypi/v/sample -->
-<!-- [packageurl]: https://pypi.org/project/sample/ -->
+<!-- [versionbadge]: https://img.shields.io/pypi/v/excelbudget -->
+<!-- [packageurl]: https://pypi.org/project/excelbudget/ -->
 [testbadge]: https://github.com/patrick-5546/excelbudget/actions/workflows/ci.yml/badge.svg
 [testfile]: https://github.com/patrick-5546/excelbudget/actions/workflows/ci.yml
 [covbadge]: https://img.shields.io/endpoint?url=https://gist.githubusercontent.com/patrick-5546/845b19d91f3d03c94677f6fae6eb414c/raw/covbadge.json
-<!-- [pyversionbadge]: https://img.shields.io/pypi/pyversions/sample -->
+<!-- [pyversionbadge]: https://img.shields.io/pypi/pyversions/excelbudget -->
 [licensebadge]: https://img.shields.io/github/license/patrick-5546/excelbudget
 [licenseurl]: https://github.com/patrick-5546/excelbudget/blob/main/LICENSE
