+++ pyproject.toml
@@ -163,17 +163,17 @@ upgrade = ["pip-tools"]
 "Say Thanks!" = "http://saythanks.io/to/example"
 "Source" = "https://github.com/patrick-5546/excelbudget/"
 
-# The following would provide a command line executable called `sample`
+# The following would provide a command line executable called `excelbudget`
 # which executes the function `main` from this package when invoked.
 [project.scripts]  # Optional
-sample = "sample:main"
+excelbudget = "excelbudget:main"
 
 # This is configuration specific to the `setuptools` build backend.
 # If you are using a different build backend, you will need to change this.
 [tool.setuptools]
 # If there are data files included in your packages that need to be
 # installed, specify them here.
-package-data = {"sample" = ["*.dat"]}
+package-data = {"excelbudget" = ["*.dat"]}
 
 [build-system]
 # These are the assumed default build requirements from pip:
